refactor(cli): split daemon/ns/doctor handlers into modules
